Don't invoke constructors when prototyping
The basic problem here is that if you extend something that takes constructor arguments (or does anything interesting in its constructor) we invoke that constructor in a bad state (no args, no useful context) during JX.install().

There are a bunch of partial solutions which try to balance concerns, but I actually think we're okay with this more extreme version: simply no-op the constructor during prototyping.

This basically just means it's your responsibility to call this.__super__.apply(…) in your constructor if you extend. I think that's fine.

All this shuts down is crazy metaprogramming BS where you define and attach methods in your constructor, and even that should work OK as long as the extenders call __super__.

We can add a hook here (instead of just "return;") in the future if we ever find a use case, but I can't really think of any that make any sense.