Drop legacy "/* -lite */" junk from Javelin sync script
Summary:
We got rid of the last of this mess when we dropped JX.Book so we can
nuke this hack.

Test Plan:
Unable to actually test this since I don't have an FB trunk anymore,
but it's a stragith patch of D45 which I probably tested in the past.

Reviewed By: tomo
Reviewers: tomo, mroch
CC: tomo
Differential Revision: 161