[JX.Request] Add "open" event
Summary:
Evan was right, it makes sense to have an event that people can listen to to
screw with the transport. Turns out
that many of the transport properties can't be set until after open() has been
called on it, so we need another
hook.

I like this approach more now than I did before because we pass the JX.Request
instance into the callback
instead of the raw XHR. The user can now call getTransport() himself.

Test Plan:
in firebug:

  var r = new JX.Request('test/');
  r.listen('open', function(req) { req.getTransport().withCredentials = true;
});
  r.send();
  r.getTransport()

and inspected the transport. withCredentials was set to true.

Reviewed By: aran
Reviewers: tomo, epriestley, aran
CC: aran, nikolay
Differential Revision: 348