fix javelin .arcconfig and tweak .divinerconfig
Summary:
Not sure if these are right, but I can't send diffs to Phabricator for
Javelin without the arcconfig change, and we just moved the repo over to
github/facebook, right?

Let me know if I'm doing it wrong, or if I should only change my
arcconfig and then git exclude it?

Test Plan:
none really

Reviewed By: epriestley
Reviewers: epriestley
CC: aran, epriestley
Differential Revision: 367