Close JX.Workflow dialogs when the user presses escape
Summary:
Allow 'escape' to trigger a click on the cancel button. This feature was removed
primarily because simpkins has workflow which involves mashing escape over and
over again to input text, and there didn't seem to be a super compelling use
case for it. There's a stronger argument now that we have keyboard shortcuts,
and a couple people expressed regret at losing it.

Test Plan:
Hit escape to close dialogs.

Reviewed By: jungejason
Reviewers: jungejason, tuomaspelkonen, aran, mroch, tomo
Commenters: mroch
CC: aran, epriestley, jungejason, mroch
Differential Revision: 414