Add JX.log in non-dev mode to prevent silent fails
Summary: JX.log is part of the public API (it is listed in the @requires
block in util.js). In non-dev mode this method gets stripped out and
could potentially cause errors that the developer overlooked. Adding
JX.log as an empty function in non-dev mode.

Test Plan: Ensure the JX.log is available if __DEV__ is false.

Reviewers: tomo, epriestley

CC:

Differential Revision: 492