Throw when alterClass is called without third arg
Summary: If no third argument is provided, this method will behave
differently depending on whether the element has the class or not.
Javelin discourages state management through className. Throwing an
error in DEV can help developers spot this mistakte.

Test Plan: Test that it throws in DEV if anything but true or false are
provided as third argument.

Reviewers: mroch, epriestley, tomo

CC:

Differential Revision: 511