Throw an error for non-existing behaviors.
Summary:
This is very defensive. This throws in prod and it can be
caught outside of initBehaviors if one or many behaviors are not
available. If it is an unimportant behavior and if there is no attempt
to catch it on the outside, this can likely be ignored by library users
if initBehaviors is called in its own execution context.

Test Plan: Specs!

Reviewers: tomo, mroch, epriestley, jg

Reviewed By: epriestley

CC: aran, epriestley

Differential Revision: 836