Client-side views
Summary:
 - A "View" is a thin wrapper around JX.$N, but permits
     - composition
     - specialization via Javelin inheritance
 - HTMLView is dumb but conveniently adapts $N to views
 - The ViewInterpreter supports a pure JS decent syntax/EDSL for these
views, akin to a slightly cleaner version of nested JX.$N. It's usable as a
target language for CoffeeScript input, too, which makes it even cleaner.
 - The view placeholder Behavior allows a server-side view (AphrontView, XHP)
to integrate with client-side views. This is demonstrated in a separate diff.

High-level goals here are to:
 - decouple client-side templating from full client-side MVC, so if we just want
the templating we don't
necessarily need to pull in a whole MVC framework
 - give an example "pretty okay" template syntax which is intended as a Proof of
Concept for my "just use javascript" philosophy

Test Plan: Specs, build a page using the template placeholder behavior

Reviewers: epriestley, cpojer, mroch, tomo

Reviewed By: epriestley

CC: aran, rzadorozny, epriestley, cpojer, mroch, jungejason

Differential Revision: 907