RFC: Order DOM event handlers by shortest matching sigil distance from target node.
Summary:
A suggested change to the order in which matching event handlers are called.
Currently javelin claims no guarantee for any kind of order, but in practice the
order is the order in which the event handlers were defined. This keeps order of
definition as the secondary sort, but makes primary sort the distance of the
matching sigil node from the event target node.

This means if you have markup:

  <div data-sigil="one">
    <div data-sigil="two">
      <div data-sigil="three">
        Click me
      </div>
    </div>
  </div>

That a handler listening to the sigil path ['three'] would execute before a
handler listening to the sigil path ['two'].

For sigil paths with more than one sigil, it will take the shortest distance to
match, so sigil path ['one', 'three'] would execute before ['two'].

For handlers which listen to all events, the distance is assumed to be "0" in
effect pointing to the target node, so listening to null would execute before
['two'] but with the same priority (in this case) as ['three'], falling back on
the order in which they were assigned.

This has the benefit of mirroring the behavior of the DOM event API's bubbling
mechanism so that a more specific region of the DOM recieves events before ta
more general region.

Test Plan: Tested against the fb m-site and tested in the console with the cases
listed in the summary.

Reviewers: epriestley, mroch, yungsters

Reviewed By: epriestley

CC: aran, epriestley, jg, hedger, mroch, leebyron, akrieger

Differential Revision: 925