RFC: Remove event listeners from the handler
Summary:
A couple months ago I submitted an RFC for JX.Stratcom.listenOnce for which we
agreed there was too few cases to support the complexity and weight that the
code added.

This is another approach to the problem that is more flexible, less complex and
has a clear API.

This allows for a handler to call JX.Stratcom.removeCurrentListener(). This
allows for the listenOnce case by having the handler always call this method,
but also allows for more complex behavior by calling only in a specific case.

Test Plan: tested by adding 3 listeners to click, the middle one calling
removeCurrentListener() in the handler. Verified all 3 were called initially,
and the removed did not get called another time where the others did.

Reviewers: mroch, tomo, epriestley, aran

Reviewed By: epriestley

CC: aran, leebyron, epriestley

Differential Revision: 957