JX.URI: Blacklist URI Characters
Summary:
**Don't allow '\' in hostnames**
Some browsers s/\\/\// in URLs, so http://yahoo.com\.facebook.com/ is
treated as http://yahoo.com/.facebook.com/. Break this by disallowing
'\' in hostnames.

**In URI, check the domain for invalid characters.**
Because browsers don't handle URIs consistently, there are a lot of ways to
trick the URI class into parsing a URI in a different way than a real browser
would. To help fix this, be more strict about the characters we accept in the
domain.

[This is a sync from the Facebook repository.]

Reviewers: epriestley, tomo, mroch, aran

Test Plan:
Run Jasmine:

  specs/runner.html

Differential Revision: 981