Create JX.now(), a wrapper for timestamp generation
Summary: The motivation for this is to give a coherent API for timestamp
generation that is also as fast as possible. IE is the notable pain point that
prevents us from just using Date.now. See
http://jsperf.com/date-initialization/2 and http://jsperf.com/property-lookups
for the raw numbers on timestamp generation. There has already been a lot of
back and forth about this diff internally, but the outcome seems to be in favor
so I'm proceeding with this. If nothing else, JX.now() is a cleaner API than
+new Date(), which is only marginally better than int casting with bitwise
operators and other JS jedi mind tricks, and is only two extra bytes.

Reviewers: yungsters, kdelong, mroch, epriestley, tomo

CC:

Test Plan: Tested within the Facebook Mobile site in my sandbox

Differential Revision: 983