JX.History: Workaround Safari "location.replace" Bug
Summary:
This works around a Safari bug that affects the history stack when using
##location.replace## and setting ##location.hash##. The fix is to use
##history.replaceState## when possible in place of ##location.replace##.

The bug can be reproduced in Safari by entering the following:

  location.hash = "A"; // Now at "A".
  location.replace(location.href + "B"); // Now at "AB".
  location.hash = "C"; // Now at "C".
  history.go(-1); // Goes to "A", but should go to "AB".

This has been fixed on the nightly Webkit build, but it affects both iOS 4 and
5.

Test Plan:
Force ##historytest.html## to install with ##JX.History.HASHCHANGE##.
Open the spec file in Safari.
Click the Safari Bug button.

Reviewers: cpojer, leebyron

Reviewed By: leebyron

CC: aran, leebyron

Differential Revision: 1006