Listen to 'input' events in Typeahead for text input changes that don't fire 'keydown' events
Summary:
CJK input completition (such as selecting the Kanji representation from the
corresponding Romaji in Japanese) changes the input field value, but doesn't
fire a keydown event.

To fix this, we listen to the 'input' event as well. Since we use Javelin for
event delegation, we won't get JS fatals due to the guards in JX.enableDispatch
and listening to events that are never fired is safe. Since we're still
listening to keydown events anyways, we should be fine if the 'input' event
isn't supported.

This might also fix auto-correct in text inputs, since auto-correct might not
fire a keydown event.

Test Plan:
Tested in Chrome and iOS Safari on Facebook's mobile typeahead
Typed 'Keito' using the Romaji keyboard - noted results
Selected the first Kanji approximation of 'Keito' - results changed to reflect
selected Kanji

Reviewers: jg, yungsters, tomo, leebyron

Reviewed By: leebyron

CC: aran, leebyron

Differential Revision: 1057