Typeahead: Simplify Root Manipulation
Summary:
When used in an ##<iframe />##, for some reason the combination of clearing the
root node and removing it from the tree causes the latest version of Mobile
Safari (bundled with iOS 5) to crash (or spin up CPU to 100%).

This diff simplifies the logic that deals with the root node; the behavior
should be the same.

Test Plan: It still... works? I'm open for suggestions on how to test this
properly besides running it on TRUNK.

Reviewers: epriestley, tomo, cpojer, leebyron

Reviewed By: epriestley

CC: aran, epriestley

Differential Revision: 1058