Fix Typeahead/Tokenizer w/ Same Hardpoint
Summary: This fixes a bug caused when a Typeahead and Tokenizer are both
constructed using the same hardpoint. The Typeahead appends a node to store
results in, but the Tokenizer destroys the contents of the hardpoint (and the
node to store results in with it).

Test Plan: Verified that a Typeahead and Tokenizer constructed using the same
hardpoint still worked on TRUNK.

Reviewers: epriestley, isaac, cpojer

CC: aran

Differential Revision: 1072