[jx] fix bug in JX.Request
Summary: JX.JSON can return null in some cases, such as when the input is empty
or not a string. In that case we can't blindly check for response.error.

Test Plan: Provide a null response to a JX.Request and it no longer complains
that we're accessing members of a null object.

Reviewers: epriestley, aran, tomo, mroch

Reviewed By: epriestley

CC: aran, epriestley, leebyron

Differential Revision: 1113