[jx] Merge some FB extensions: set/addData, start event, abortable events
Summary:
A couple months back this was added to the FB checkout of Javelin and wasn't
submitted back into the open source repo. I've made a few additions to those
additions (fb peeps: https://phabricator.fb.com/D272435).

A repeat of the rationale for these additions:

A start event allows for a global listener to get a reference to the request to
pre-process it or validate it. We use this in a few places on the mobile site.
It's useful to have a hook before the transport is assembled. In order to do
validation, you need to be able to abort the request. This also checks to ensure
that if an abort was called that the send() method will bail. I've added a check
to ensure the event itself wasn't prevented as well.

Test Plan: JSLint, a few months of battle.

Reviewers: epriestley, mroch, tomo, yungsters, jg

Reviewed By: epriestley

CC: aran, epriestley, leebyron

Differential Revision: 1168