[jx] allow mergeData() to be called multiple times.
Summary:
We're trying to get early flush pages to work and ran into an issue with javelin
metadata where we wish to append meta data. For example:

  <div data-meta="0_0"> first thing </div>
  <script> JX.Stratcom.mergeData(0, {'0' : data}); </script>

  [flush]

  <div data-meta="0_1"> second thing </div>
  <script> JX.Stratcom.mergeData(0, {'1' : data}); </script>;

Calling this as is will overwrite the previously merged data (in the case of
block 0, will also try to start the event queue again).

We can't allocate a new block for each flush due to possible race conditions.
For example, an async request could be dispatched before the page has completed
flushing, at which point the flush and async request could possibly be using the
same block.

The proposed solution is to allow mergeData to be called multiple times for the
same block id, appending data each time. In __DEV__ we check to ensure that
previously merged keys are not overwritten. On the server side, this is easy to
implement by doing a small amount of cleanup after each flush.

Existing behavior is the same with the exception of now throwing in __DEV__
should you attempt to merge into the same block id naively. IMO this exception
should have existed in the first place.

Test Plan: unit test. Also tested in our flushed page environment.

Reviewers: epriestley, yungsters, tomo, mroch, jg

CC: aran

Differential Revision: 1177