Provide a 'change' event to JX.Tokenizer
Summary: Provide an event like a 'change' event for <select />. The concrete
motivation for this is that we want to update the preview panel in Differential
and Maniphest when a user is executing an "Add Reviewers..." or "Add CCs..."
action and changes the users they're adding.

Test Plan: Installed a JX.Tokenizer listener for 'change', got events when I
altered the value of the tokenizer.

Reviewers: vrana, cpojer, aran, tomo, mroch

Reviewed By: cpojer

CC: aran, epriestley, cpojer

Differential Revision: https://secure.phabricator.com/D1280