Improve usability in Tokenizer
Summary:
Currently when text is not resolved to a token in Tokenizer.js
the user is not aware their attended action was not completed.
Example if you add a user to a diff but misspell their name, and
quickly submit the form. You might not notice that the
token was not resolved, and your reviewer was not added to your diff.

My change will remove the text you typed if the token doesn't
resolve successfully. Similar to what facebook does in the new
message dialog box when adding friends.

@epriestley Sorry I had some git issues and needed to abandon D1358.

Test Plan:
Tested all use cases in differential for adding reviews.
-successfully adding a resolved token user
-typing text that doesn't resolve and hitting enter
-typing text that doesn't resolve and clicking on a different field
-typing text that doesn't resolve and tabbing out of the input
I also tested the above actions in the project edit screen.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran, epriestley

Differential Revision: https://secure.phabricator.com/D1371