[JX] Ensure empty string is handled correctly.
Summary: It's possible for the server to return an {__html:''} in which case
this current logic fails. This explicitly checks against null/undefined instead
of falsey in order to allow empty string to pass.

Test Plan: Continues to work as expected. Server returning emtpy string html
blob now works as expected.

Reviewers: epriestley, cpojer, aran, tomo

Reviewed By: epriestley

CC: aran, epriestley

Differential Revision: https://secure.phabricator.com/D1393