Fix up/down in Tokenizers
Summary:
D1371 added an overbroad rule here which we actually don't need.

The intent was to clear the form if you hit "return" with a partial token, but
that's actually sort of bad. The overall goal here (making it more clear when
something isn't a token) is well-served by just the clear-on-blur part.

Remove the problematic, overbroad rule.

Test Plan:
  - Hit up/down on a tokenizer to choose between tokens, it worked.
  - Hit "return" with an invalid token, token was unaffected. This is fine.

Reviewers: vrana, btrahan

Reviewed By: vrana

CC: aran, epriestley

Maniphest Tasks: T927

Differential Revision: https://secure.phabricator.com/D1749