[javelinsymbols] -- Add support to javelinsymbols for JX.behavior()
Summary:
Currently javelinsymbols can detect the installation of a class...but not behaviors. This adds support to the javelinsymbols processor to identify behavior-registrations and specify them using a new '*' directive.

I plan to use javelinsymbols for a project to implicitly identify a unique name for a given JS resource based on the JX.install()s and JX.behavior()s calls that exist in the file.

Test Plan: Compiled and ran javlinesymbols against a file with a JX.install(), a file with JX.behavior(), and a file with both to verify the output was as expected.

Reviewers: leebyron, epriestley, jg

Reviewed By: epriestley

CC: aran, epriestley

Differential Revision: https://secure.phabricator.com/D2023