Typeahead: separate the name field from the field we complete on
Summary:
Let typeahead datasources provide an optional field: tokenizable.  If present,
the typeahead tokenizes and completes on this field instead of name, but still
uses the name when inputting data into the <input>.

For example, I can have an input for Phabricator commit identifiers, and build
typeahead from (commit-identifier, commit-message-summary-line) pairs.  The
'display' field formats both prettily, the 'name' is the commit-identifier, and
the 'tokenizable' contains both commit-identifier and commit-message
concatenated.  The typeahead lets me search both, but the resulting value in
the <input> is just the commit-identifier.

Test Plan:
Made an example as described above which queries all my commits to a
repo from the past week.

Reviewers: epriestley

Reviewed By: epriestley

CC: aran

Differential Revision: https://secure.phabricator.com/D2070