[JX.Request] Add toggle for the expectation of a CSRF guard in the response
Summary:
I am working on a project that needs to send out an XHR to the FB platform API.

The FB API doesn't return responses that are pre-pended with a CSRF guard, so I need to be able to toggle the expectation of that on the request object.

Test Plan: Copied this change into the code that tries to execute this API request an saw things work with and without toggling .setExpectCSRFGuard().

Reviewers: epriestley, leebyron, al

Reviewed By: epriestley

CC: aran, Koolvin

Differential Revision: https://secure.phabricator.com/D2573