Fix TypeaheadSource not considering excludeids
Summary:
Previously, TypeaheadSource doesn't consider the excludeids
that are updated after the datasource has been created. It should also
check for it every time the match results is called so that it supports adding/removing exclude ids on lookup time.

Test Plan:
Try to insert an exclude id via javascript after datasource
has been initialized. Type in the name, make sure the name is now not included in the typeahead

Reviewers: linjeffrey, jeffmo, epriestley

Reviewed By: epriestley

CC: aran, epriestley

Differential Revision: https://secure.phabricator.com/D3108