Fix an issue where the the internal state of a Typeahead was not properly cleared after a token was selected
Summary:
JX.Typeahead keeps a copy of the input state so we can discard keypresses which don't change the input, but we were manually clearing the input rather than calling clear() properly. Additionally, we don't clear the internal state in clear().

Instead, call clear() from JX.Tokenizer, and reset the internal state in clear().

A specific way to trigger this bug is type "z" in a Tokenizer (assuming it matches 2+ tokens), get a match, hit return to select it, and then type "z" again. The incorrect behavior (before this patch) was for the typeahead suggestion list to remain hidden until you typed another character. The correct behavior (after this patch) is for the suggestion list to reopen.

Test Plan: Typed "z", selected a token, typed "z", got a suggestion list.

Reviewers: hendri

Reviewed By: hendri

CC: jeffmo, linjeffrey, aran

Differential Revision: https://secure.phabricator.com/D3159