Add a getResult(id) method to TypeaheadSource
Summary:
In some places, like the `choose` event of JX.Typeahead and the `getTokens()` method of JX.Tokenizer, we return result IDs and some piece of derived data (the node in the former case, the display string in the latter case). In most cases, these are sufficient, if somewhat awkward and possibly poor architectural decisions.

However, sometimes we need the raw result with any additional data it may carry (see D3143). Add a method to do a lookup by ID.

Test Plan: Built an example using this, see D3143.

Reviewers: hendri

Reviewed By: hendri

CC: jeffmo, linjeffrey, aran

Differential Revision: https://secure.phabricator.com/D3161