JX.Request: Support raw data instead of file
Summary:
http://www.w3.org/TR/XMLHttpRequest2/#the-send-method

The XMLHttpRequest spec allows ##send## to take any of the following:
- ArrayBuffer
- Blob
- Document
- string
- FormData

For example, using FormData allows the caller to send files along with
data. This ends up being a straight rename of the property and
deprecating the file property.

Test Plan:
  var request = new JX.Request('someurl', function() { debugger; });
  var formData = new FormData();
  formData.append('foo', 'bar');
  request
    .setRawData(formData)
    .send();

See that formData is sent as part of the request. (check in web inspector)

Reviewers: epriestley, yungsters, tomo

Reviewed By: epriestley

CC: aran, Korvin

Differential Revision: https://secure.phabricator.com/D3727